+++ lib/github.mli
@@ -23,6 +23,7 @@ end
 (** {1 Publish} *)
 
 val publish_distrib :
+  ?token:Fpath.t ->
   dry_run:bool ->
   msg:string ->
   archive:Fpath.t ->